requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.7.0
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gotrue==2.9.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.26.4
packaging==24.1
postgrest==0.17.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
realtime==2.0.5
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
storage3==0.8.0
StrEnum==0.4.15
supabase==2.8.1
supafunc==0.6.0
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.1
websockets==13.1
Werkzeug==3.0.2
yarl==1.13.1
gunicorn